api/requirements.txt
Flask
Flask-Cors
boto3
psycopg2-binary
python-dotenv